Core DW Infrastructure/serverpage/requirements.txt
streamlit
msal
pyjwt
cryptography
psutil
